templates/gan/requirements.txt
torch>=1.7.0
pytorch-ignite>=0.4.4
torchvision>=0.8.0
matplotlib>=3.3.0
pandas
pytest
{{ handler_deps }}
{{ logger_deps }}
